+++ iso20022-core/src/main/java/com/prowidesoftware/swift/model/mx/AppHdrType.java
@@ -24,7 +24,8 @@ public enum AppHdrType {
     LEGACY(LegacyAppHdr.NAMESPACE, LegacyAppHdr.class),
     BAH_V1(BusinessAppHdrV01.NAMESPACE, BusinessAppHdrV01.class),
     BAH_V2(BusinessAppHdrV02.NAMESPACE, BusinessAppHdrV02.class),
-    BAH_V3(BusinessAppHdrV03.NAMESPACE, BusinessAppHdrV03.class);
+    BAH_V3(BusinessAppHdrV03.NAMESPACE, BusinessAppHdrV03.class),
+    BAH_V4(BusinessAppHdrV04.NAMESPACE, BusinessAppHdrV04.class);
 
     private String namespace;
     private Class headerClass;